Move Mongoose to a peer dependency
Mongoose uses an internal global cache of data for schemas etc. A consumer of this package will make the connection to MongoDB using the version of Mongoose installed in the parent package. Move Mongoose to a peer dependency so this package and its package use the same Mongoose package.

Also add Mongoose as a dev dependency for local development. Dev dependencies are not installed when this package is required by another.